Normative: Unify Intl.Locale construction with option resolution
This patch rewrites the Intl.Locale constructor to be more similar
to other Intl constructors, being based on a central call to
ResolveLocale. The biggest observable change is that invalid
options, both in Unicode extension tags and in the options bag,
are usually ignored, rather than throwing an exception.
If throwing an exception is important, this could be recovered
in a follow-on change without undoing the common factoring.