Normative: Use locale internal slots for construction
When a Intl object (e.g., a formatter) is constructed, if an
Intl.Locale object is passed as a parameter, treat the Locale as
it is represented in internal slots, rather than invoking a
(potentially modified) toString method. This patch is written
in terms of converting the Locale to a String in a fixed way,
before it gets reparsed; the specification would probably
benefit from a follow-on editorial change to refactor ResolveLocale
to deal in terms of Locale objects instead.

Closes #9