Editorial: Layer locale processing under Intl.Locale
This patch consolidates the Intl specification by unifying the
places where language tag strings are parsed and produced: All
processing is moved to the core of the Intl.Locale constructor.
This

Options processing is still left not unified between Intl.Locale and
other Intl constructors. This is because the semantics of other
Intl constructors are significantly different from Intl.Locale
in multiple ways:
- Unicode tags such as "ca" fall back to locale-specific defaults
- Tags are checked for presence of locale data, falling back to
  another value if not data is not present
- resolvedOptions() has a locale property which expresses the part
  of the given language tag that was used. It's unclear whether
  changing this would risk compatibility.

Closes #31