Normative: Cite UTS #35 for canonicalizing Unicode extension tags
Previously, the algorithm was defined in terms of direct algorithms,
based on a reading of RFC 6067. It turns out that RFC is a bit out
of date, and a more current algorithm is found in UTS #35. Rather
than copying that algorith here, this patch simply provides a
normative reference from one standard to another.

Closes #43