Normative: Define error recovery behavior for likely subtags
When Intl.Locale.prototype.maximize() and minimize() are called on
locales which don't appear in CLDR's database of known locales with
likely subtags, the semantics were previously unclear. With this
patch, a new, identical Locale object is returned.

These semantics encode the conclusion of the discussion in the
June 2018 Intl meeting.

Closes #42