Remove the call to CanonicalizeLanguageTag in InsertUnicodeExtension
InsertUnicodeExtension only called CanonicalizeLanguageTag to maximize the shareable number of steps with 9.2.6 ResolveLocale, so that when the Intl.Locale proposal is merged into ECMA-402, ApplyUnicodeExtensionToTag and ResolveLocale can both call InsertUnicodeExtension to insert a Unicode extension into a language tag. And ResolveLocale only calls CanonicalizeLanguageTag so that implementations which implement the optional steps in 6.2.3 CanonicalizeLanguageTag, namely to apply canonicalization of the Unicode extension per RFC 6067, are allowed to use this extension canonicalization after adding the new Unicode extension in ResolveLocale.
But now that Intl.Locale has made the RFC 6067 / UTR 35 canonicalization normative, it no longer makes sense to canonicalize the already canonicalized language tag with an also already canonicalized Unicode extension another time.